+++ services/favorite.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 )
 
-func GetUserFavorites(userID uint) ([]uint, error) {
+func GetUserFavorites(userID int64) ([]int64, error) {
 	favorites, err := userRepo.GetFavorites(userID)
 	if err != nil {
 		logger.Println(err)
@@ -13,7 +13,7 @@ func GetUserFavorites(userID uint) ([]uint, error) {
 	return favorites, nil
 }
 
-func CheckFavorite(userID uint, videoID uint) bool {
+func CheckFavorite(userID int64, videoID int64) bool {
 	favorite, err := userRepo.CheckFavorite(userID, videoID)
 	if err != nil {
 		logger.Println(err)
@@ -22,7 +22,7 @@ func CheckFavorite(userID uint, videoID uint) bool {
 	return favorite
 }
 
-func Favorite(userID uint, videoID uint) error {
+func Favorite(userID int64, videoID int64) error {
 	if err := userRepo.AddFavorite(userID, videoID); err != nil {
 		logger.Println(err)
 		return errors.New("收藏失败")
@@ -30,7 +30,7 @@ func Favorite(userID uint, videoID uint) error {
 	return nil
 }
 
-func Unfavorite(userID uint, videoID uint) error {
+func Unfavorite(userID int64, videoID int64) error {
 	if err := userRepo.DeleteFavorite(userID, videoID); err != nil {
 		logger.Println(err)
 		return errors.New("取消收藏失败")
